added rotation tuning script

+++ examples/usage_demos/l298n_speed/l298n_speed.ino
@@ -39,12 +39,12 @@ void l298MotorBrake()
 }
 
 // Wrapper functions for the encoder
-int encoderReadFunc()
+long encoderReadFunc()
 {
     return myEncoder.read();
 }
 
-void encoderWriteFunc(int newPosition)
+void encoderWriteFunc(long newPosition)
 {
     myEncoder.write(newPosition);
 }
@@ -117,7 +117,7 @@ void loop()
         }
         else if (command.startsWith("MOVE="))
         {
-            int newPosition = command.substring(5).toInt();
+            long newPosition = command.substring(5).toInt();
             if (newPosition != 0)
             {
                 servo.move(newPosition);
@@ -131,7 +131,7 @@ void loop()
         }
         else if (command.startsWith("MAXPWM="))
         {
-            int maxPWM = command.substring(7).toInt();
+            long maxPWM = command.substring(7).toInt();
             if (maxPWM >= 0 && maxPWM <= 255)
             {
                 servo.setMaxPWM(maxPWM);
